Add files via upload
Fixed the addpath in all example livescripts. The paths were modified assuming that the root folder is the mTRF-toolbox folder (as opposed to the example folder, where the scripts are stored). That seems to have changed in Matlab Online since the previous mTRF-Toolbox release. The second issue was that the addpath seems to be case sensitive for folders (while that did not seem to be an issue before). this might have something to do with recent updated with Matlab Online.